user-service/requirements.txt
# ----------------------------
# ✅ Core Framework
# ----------------------------
fastapi==0.115.0
uvicorn[standard]==0.30.3

# ----------------------------
# ✅ Database and ORM
# ----------------------------
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2

# ----------------------------
# ✅ Pydantic and Settings (v2)
# ----------------------------
pydantic==2.7.4
pydantic-settings==2.2.1
email-validator==2.2.0
python-dotenv==1.0.1

# ----------------------------
# ✅ Authentication & Security
# ----------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
secrets==1.0.2

# ----------------------------
# ✅ Caching / Rate Limiting / Queues
# ----------------------------
redis==5.0.3
aioredis==2.0.1

# ----------------------------
# ✅ HTTP Clients (for internal service communication)
# ----------------------------
httpx==0.27.0
requests==2.32.3

# ----------------------------
# ✅ Logging & Monitoring
# ----------------------------
loguru==0.7.2
structlog==24.1.0

# ----------------------------
# ✅ Serialization / Utilities
# ----------------------------
orjson==3.10.7
python-dateutil==2.9.0
tenacity==9.0.0

# ----------------------------
# ✅ Testing & Development Tools
# ----------------------------
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
black==24.4.2
flake8==7.0.0
isort==5.13.2

# ----------------------------
# ✅ Optional: Celery / Background Jobs (for later)
# ----------------------------
celery==5.4.0
kombu==5.4.2

# ----------------------------
# ✅ Optional: Observability / Metrics (future use)
# ----------------------------
prometheus-client==0.20.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
